Add bin/vim.wrapper base launcher
I use this wrapper to launch vim in different configurations. Whenever I
have a vim.* file in my vim-path, I create a symlink to this script in
my dot.vim/bin/ directory as a launcher for this config.

vim.bigfile -> starts with ~/.vim/vim.bigfile and so on.

Also vim.blank is available as a hardcoded default for a blank vim
without custom configuration.